refactor: separate type models from User
